Use conditional rendering instead of navigations
