+++ src/screens/LoginScreen.tsx
@@ -1,18 +1,21 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { TouchableOpacity, View } from 'react-native';
 import PropTypes from 'prop-types';
 
 import { signInWithEmailAndPassword } from 'firebase/auth';
 import { useAuthState } from 'react-firebase-hooks/auth';
 
+import { createStackNavigator } from '@react-navigation/stack';
+
 import { auth } from '../../firebase';
+
+import SignUpScreen from './SignUpScreen';
 
 import Button from '../components/Button';
 import Input from '../components/Input';
 import Text from '../components/Text';
 
-import { colors } from '../styles/styles';
-import styles from '../styles/FriendsScreen.styles';
+import { colors, globalStyles } from '../styles/styles';
 
 const login = (email: string, password: string) => {
   signInWithEmailAndPassword(auth, email, password)
@@ -27,24 +30,17 @@ interface Props {
     goBack: () => {}
   },
 }
-
-export default function LoginScreen({ navigation }: Props) {
+function LoginPage({ navigation }: Props) {
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
-  const [user, loading, error] = useAuthState(auth);
+  const [, , error] = useAuthState(auth);
 
   if (error) {
     console.log(error);
   }
 
-  useEffect(() => {
-    if (user && !loading) {
-      navigation.goBack();
-    }
-  }, [user, loading]);
-
   return (
-    <View style={styles.main}>
+    <View style={globalStyles.centered}>
       <Input
         placeholder="Email"
         onChangeText={setEmail}
@@ -67,9 +63,24 @@ export default function LoginScreen({ navigation }: Props) {
   );
 }
 
-LoginScreen.propTypes = {
+LoginPage.propTypes = {
   navigation: PropTypes.shape({
     navigate: PropTypes.func.isRequired,
     goBack: PropTypes.func.isRequired,
   }).isRequired,
 };
+
+const RootStack = createStackNavigator();
+
+export default function LoginScreen() {
+  return (
+    <RootStack.Navigator>
+      <RootStack.Group screenOptions={{ headerShown: false }}>
+        <RootStack.Screen name="Login" component={LoginPage} />
+      </RootStack.Group>
+      <RootStack.Group>
+        <RootStack.Screen name="Sign Up" component={SignUpScreen} />
+      </RootStack.Group>
+    </RootStack.Navigator>
+  );
+}
